+++ package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "node-sass": "^9.0.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0"
   },
   "lint-staged": {
     "src/**/*.{ts,tsx}": [
