+++ src/pages/Error/Error.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Error: React.FC = () => {
+  return <> Error Page </>;
+};
+
+export default Error;
