+++ src/pages/Home/Home.module.scss
@@ -1,4 +1,4 @@
-@import "./styles/variables.scss";
+@import "../../styles/variables.scss";
 
 body {
   margin: 0;
